Fixed #22979 Move bug* tests
