+++ tests/admin_autodiscover/admin.py
