+++ tests/admin_autodiscover/models.py
