+++ tests/admin_autodiscover/tests.py
@@ -3,12 +3,12 @@ from unittest import TestCase
 from django.contrib import admin
 
 
-class Bug8245Test(TestCase):
+class AdminAutoDiscoverTests(TestCase):
     """
     Test for bug #8245 - don't raise an AlreadyRegistered exception when using
     autodiscover() and an admin.py module contains an error.
     """
-    def test_bug_8245(self):
+    def test_double_call_autodiscover(self):
         # The first time autodiscover is called, we should get our real error.
         with self.assertRaises(Exception) as cm:
             admin.autodiscover()
